localore/requirements.txt
Django>=1.8,<1.9
django_environ==0.4.0
django_localflavor==1.2.0
django_overextends==0.4.0
elasticsearch>=1.0.0,<2.0.0
embedly==0.5.0
psycopg2==2.6.1
pytz
wagtail==1.3.1
wagtailmodeladmin==2.4.8
Wand==0.4.1
-e git+https://github.com/django-compressor/django-compressor.git@de3afb17ab102430e9d3c97034f6016c6a0d3037#egg=django-compressor
